Move and rename boss effects file
