+++ TombEngine/Game/Lara/lara_fire.cpp
@@ -19,11 +19,11 @@
 #include "Game/misc.h"
 #include "Game/savegame.h"
 #include "Math/Math.h"
+#include "Objects/Effects/Boss.h"
 #include "Objects/Generic/Object/burning_torch.h"
 #include "Objects/Generic/Object/objects.h"
 #include "Objects/ScriptInterfaceObjectsHandler.h"
 #include "Objects/TR3/Entity/PunaBoss.h"
-#include "Objects/TR3/Object/tr3_boss_object.h"
 #include "ScriptInterfaceGame.h"
 #include "ScriptInterfaceLevel.h"
 #include "Sound/sound.h"
