+++ TombEngine/Objects/Effects/Boss.h
